+++ .gitignore
@@ -7,7 +7,6 @@
 *.out
 crawlera-headless-proxy
 vendor/
-ccbuilds/
 
 # autogenerated
 certs.go
