+++ go.mod
@@ -5,21 +5,14 @@ require (
 	github.com/BurntSushi/toml v0.3.1
 	github.com/alecthomas/template v0.0.0-20160405071501-a0175ee3bccc // indirect
 	github.com/alecthomas/units v0.0.0-20151022065526-2efee857e7cf // indirect
-	github.com/elazarl/goproxy v0.0.0-20181111060418-2ce16c963a8a
 	github.com/go-chi/chi v3.3.3+incompatible
-	github.com/hashicorp/go-uuid v1.0.0
 	github.com/juju/errors v0.0.0-20181118221551-089d3ea4e4d5
 	github.com/karlseguin/ccache v2.0.3-0.20181126132943-142396791ef4+incompatible
-	github.com/kr/pty v1.1.3 // indirect
 	github.com/montanaflynn/stats v0.0.0-20180911141734-db72e6cae808
-	github.com/nbio/st v0.0.0-20140626010706-e9e8d9816f32 // indirect
 	github.com/pmezard/adblock v0.0.0-20171028110701-edfb97ad89cd
 	github.com/sirupsen/logrus v1.2.0
-	github.com/stretchr/testify v1.2.2
-	github.com/valyala/fasthttp v1.0.0
 	golang.org/x/crypto v0.0.0-20181203042331-505ab145d0a9 // indirect
-	golang.org/x/sys v0.0.0-20181210030007-2a47403f2ae5 // indirect
+	golang.org/x/sys v0.0.0-20181211161752-7da8ea5c8182 // indirect
 	golang.org/x/text v0.3.0 // indirect
 	gopkg.in/alecthomas/kingpin.v2 v2.2.6
-	gopkg.in/h2non/gock.v1 v1.0.12
 )
