+++ layers/sessions.go
@@ -4,16 +4,9 @@ import (
 	"sync"
 
 	"github.com/9seconds/httransform"
-	"github.com/karlseguin/ccache"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/scrapinghub/crawlera-headless-proxy/config"
-)
-
-const (
-	sessionsChansMaxSize      = 100
-	sessionsChansItemsToPrune = sessionsChansMaxSize / 2
-	sessionsChansTTL          = sessionClientTimeout * 2
 )
 
 type SessionsLayer struct {
@@ -21,7 +14,6 @@ type SessionsLayer struct {
 	crawleraHost string
 	crawleraPort int
 	clients      *sync.Map
-	sessionChans *ccache.Cache
 	executor     httransform.Executor
 }
 
@@ -39,14 +31,15 @@ func (s *SessionsLayer) OnRequest(state *httransform.LayerState) error {
 		state.RequestHeaders.SetString("X-Crawlera-Session", value)
 	case chan<- string:
 		state.RequestHeaders.SetString("X-Crawlera-Session", "create")
-		s.sessionChans.Set(clientID, value, sessionsChansTTL)
+		state.Set(sessionChanContextType, value)
 	}
 
 	return nil
 }
 
 func (s *SessionsLayer) OnResponse(state *httransform.LayerState, err error) {
-	if err != nil {
+	if channelUntyped, ok := state.Get(sessionChanContextType); ok && err != nil {
+		close(channelUntyped.(chan<- string))
 		return
 	}
 
@@ -60,18 +53,16 @@ func (s *SessionsLayer) OnResponse(state *httransform.LayerState, err error) {
 }
 
 func (s *SessionsLayer) onResponseOK(state *httransform.LayerState) {
-	if item := s.sessionChans.Get(getClientID(state)); item != nil {
-		channel := item.Value().(chan<- string)
-		defer close(channel)
+	if channelUntyped, ok := state.Get(sessionChanContextType); ok {
+		sessionID, _ := state.ResponseHeaders.GetString("x-crawlera-session")
+		channelUntyped.(chan<- string) <- sessionID
+		close(channelUntyped.(chan<- string))
 
-		if !item.Expired() {
-			sessionID, _ := state.ResponseHeaders.GetString("x-crawlera-session")
-			channel <- sessionID
-			getMetrics(state).NewSessionCreated()
-			getLogger(state).WithFields(log.Fields{
-				"session-id": sessionID,
-			}).Info("Initialized new session")
-		}
+		getMetrics(state).NewSessionCreated()
+
+		getLogger(state).WithFields(log.Fields{
+			"session-id": sessionID,
+		}).Info("Initialized new session")
 	}
 }
 
@@ -80,12 +71,12 @@ func (s *SessionsLayer) onResponseError(state *httransform.LayerState) {
 	mgrRaw, _ := s.clients.Load(clientID)
 	mgr := mgrRaw.(*sessionManager)
 
+	if channelUntyped, ok := state.Get(sessionChanContextType); ok {
+		close(channelUntyped.(chan<- string))
+	}
+
 	brokenSessionID, _ := state.ResponseHeaders.GetString("x-crawlera-session")
 	mgr.getBrokenSessionChan() <- brokenSessionID
-
-	if item := s.sessionChans.Get(clientID); item != nil {
-		close(item.Value().(chan<- string))
-	}
 
 	switch value := mgr.getSessionID(true).(type) {
 	case chan<- string:
@@ -107,8 +98,8 @@ func (s *SessionsLayer) onResponseErrorRetryCreateSession(state *httransform.Lay
 		return
 	}
 
-	sessionID := state.Response.Header.Peek("X-Crawlera-Session")
-	channel <- string(sessionID)
+	sessionID := string(state.Response.Header.Peek("X-Crawlera-Session"))
+	channel <- sessionID
 	getMetrics(state).NewSessionCreated()
 
 	logger.WithFields(log.Fields{
@@ -138,7 +129,7 @@ func (s *SessionsLayer) executeRequest(state *httransform.LayerState) {
 	s.executor(state)
 
 	state.ResponseHeaders.Clear()
-	httransform.ParseHeaders(state.ResponseHeaders, state.Response.Header.Header())
+	httransform.ParseHeaders(state.ResponseHeaders, state.Response.Header.Header()) // nolint: errcheck
 }
 
 func NewSessionsLayer(conf *config.Config, executor httransform.Executor) httransform.Layer {
@@ -147,7 +138,6 @@ func NewSessionsLayer(conf *config.Config, executor httransform.Executor) httran
 		crawleraPort: conf.CrawleraPort,
 		apiKey:       conf.APIKey,
 		clients:      &sync.Map{},
-		sessionChans: ccache.New(ccache.Configure().MaxSize(sessionsChansMaxSize).ItemsToPrune(sessionsChansItemsToPrune)),
 		executor:     executor,
 	}
 }
